Reorg evm utils
