+++ utils/xmtpRN/client.ts
@@ -21,7 +21,7 @@ import { CoinbaseMessagingPaymentCodec } from "./contentTypes/coinbasePayment";
 import { getXmtpClient } from "./sync";
 import config from "../../config";
 import { getDbDirectory } from "../../data/db";
-import { getCleanAddress } from "../eth";
+import { getCleanAddress } from "../evm/address";
 
 const env = config.xmtpEnv as "dev" | "production" | "local";
