+++ utils/xmtpRN/conversations.web.ts
@@ -8,7 +8,7 @@ import { saveConversations } from "../../data/helpers/conversations/upsertConver
 import { getChatStore, getSettingsStore } from "../../data/store/accountsStore";
 import { XmtpConversation } from "../../data/store/chatStore";
 import { SettingsStoreType } from "../../data/store/settingsStore";
-import { getCleanAddress } from "../eth";
+import { getCleanAddress } from "../evm/address";
 
 const protocolConversationToStateConversation = (
   conversation: Conversation
